+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Heroku deployment support with `jhipster heroku` sub-generator
 - LICENSE file added and project licensed under GPL version 3
 - Setup file for Travis CI
+- Overview with `react-big-calendar` displaying contract ends in a calendar view
 
 ### Changed